+++ tests/parse/deserializeProperty.spec.ts
@@ -1,1 +1,13 @@
+import { deserializeProperty, Property } from '../../src'
+
 // TODO
+it('unfolds lines', () => {
+    const serialized = `SUMMARY:my \r
+ long\r
+ line wowowow`
+
+    const property = deserializeProperty(serialized)
+
+    const expected = new Property('SUMMARY', 'my longline wowowow')
+    expect(property).toStrictEqual(expected)
+})